Completed spring app up to week 7
