+++ src/main/java/com/edwinsoto/dao/PersonDAO.java
@@ -1,0 +1,23 @@
+package com.edwinsoto.dao;
+
+import com.edwinsoto.adopter.Adopter;
+
+import java.util.Collection;
+import java.util.List;
+public interface PersonDAO {
+
+    Adopter insert(Adopter person);
+
+    boolean update(Adopter person);
+
+    boolean delete(int id);
+
+    Adopter findByID(int id);
+
+    boolean findByName(String name);
+
+    Collection<Adopter> naturalSort();
+
+    List<Adopter> findAll();
+
+}
