--- src/main/java/com/expeditors/AdoptionAppConfig.java
@@ -1,12 +1,0 @@
-package com.expeditors;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-
-@Configuration
-@ComponentScan({"com.expeditors"})
-public class AdoptionAppConfig {
-
-}
-
