--- src/main/java/com/expeditors/Main.java
@@ -1,107 +1,0 @@
-package com.expeditors;
-
-import com.expeditors.adopter.AdopterImpl;
-import com.expeditors.adopter.Adopter;
-import com.expeditors.animal.Animal;
-import com.expeditors.animal.Cat;
-import com.expeditors.animal.Dog;
-import com.expeditors.animal.Turtle;
-import com.expeditors.service.AdopterService;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public class Main {
-
-
-    public void main(String[] args) {
-        Main app = new Main();
-        app.run();
-    }
-
-    public void run() {
-
-        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
-        context.getEnvironment().setActiveProfiles("dev");
-        context.scan("com.expeditors");
-        context.refresh();
-
-        AdopterService adopterService = context.getBean("personService", AdopterService.class);
-
-        Adopter person1 = new AdopterImpl("Edwin Soto", "843-123-7981", "edwin.soto@email.com", true);
-        Adopter person2 = new AdopterImpl("Mackenzie Soto", "843-123-1111", "mackenzie.soto@email.com", true);
-        Adopter person3 = new AdopterImpl("Liam Soto", "123-123-3456", "liam.soto@email.com");
-        Adopter person4 = new AdopterImpl("Leila Soto", "123-123-3456", "leila.soto@email.com");
-        Adopter person5 = new AdopterImpl("Levi Soto", "123-123-3456", "levi.soto@email.com");
-
-        //Add Persons to service
-        adopterService.insertPerson(person1);
-        adopterService.insertPerson(person2);
-        adopterService.insertPerson(person3);
-        adopterService.insertPerson(person4);
-        adopterService.insertPerson(person5);
-
-
-        List<Adopter> persons = adopterService.getAllPersons();
-        int sizeListPerson = persons.size();
-        System.out.println("All Adopters:");
-        System.out.println(STR."Size of Persons List: \{sizeListPerson}");
-        persons.forEach(System.out::println);
-
-        String newline = System.lineSeparator();
-        System.out.println(newline);
-        System.out.println(STR."\{person1.getName()}, \{person2.getName()}, \{person3.getName()}, \{person4.getName()} and \{person5.getName()} are at the adoption facility");
-
-        // Add Animals to be Adopted
-        Animal cat = new Cat(1,"Fred", LocalDate.of(2024, 1, 23),  "Siamese");
-        Animal cat2 = new Cat(4, "Fido", LocalDate.of(2023, 7, 29), "persian");
-        Animal dog = new Dog(3, "Gracie", LocalDate.of(2022, 1, 23), "Terrier Mix");
-        Animal dog2 = new Dog(2, "Spot", LocalDate.of(2023,12,24),  "German Malinois");
-        Animal turtle = new Turtle(3, "Slimy", LocalDate.of(2022, 3, 23),  "Yellow Belly Slider");
-
-        // Animals make greetings
-        System.out.println();
-        cat.animalGreeting();
-        cat2.animalGreeting();
-        dog.animalGreeting();
-        dog2.animalGreeting();
-        turtle.animalGreeting();
-
-        // Set Adopted Animals
-        System.out.println(newline);
-        person1.setAdoptedAnimals(dog);
-        person2.setAdoptedAnimals(cat);
-        person2.setAdoptedAnimals(turtle);
-
-        // Get Owners of Adopted Animals
-        System.out.println(newline);
-        System.out.println(STR."\{dog.getAnimalName()}'s owner is \{dog.getOwner().getName()} adopted on \{dog.getAdoptedDate()}.");
-        System.out.println(STR."\{cat.getAnimalName()}'s owner is \{cat.getOwner().getName()} adopted on \\{dog.getAdoptedDate()}.");
-        System.out.println(STR."\{turtle.getAnimalName()}'s owner is \{turtle.getOwner().getName()} adopted on \\{dog.getAdoptedDate()}.");
-
-        System.out.println(newline);
-        System.out.println("Adopted Animals for Person2");
-        List<Animal> adoptedAnimals = person2.getAdoptedAnimals();
-        for (Animal animal: adoptedAnimals){
-            System.out.println(STR."Person 2 has adopted: \{animal.toString()}");
-        }
-
-        // Get All Persons from Service
-        System.out.println(newline);
-        System.out.println("Post Adoptions");
-        persons = adopterService.getAllPersons();
-        System.out.println(STR."Persons Size: \{persons.size()}");
-        persons.forEach(System.out::println);
-
-        // Find Adopter By Name
-        System.out.println(newline);
-        String nameToSearch = "Edwin Soto";
-        boolean isPerson = adopterService.getPersonName(nameToSearch);
-        System.out.println(STR."Person: \{nameToSearch} is \{isPerson}");
-
-        String nameToSearch2 = "Gracie Soto";
-        boolean isPerson2 = adopterService.getPersonName(nameToSearch2);
-        System.out.println(STR."Person: \{nameToSearch2} is \{isPerson2}");
-    }
-}
