+++ src/main/resources/application.properties
@@ -1,0 +1,5 @@
+spring.application.name=AdoptionApp
+
+#Profiles:
+spring.profiles.active = dev
+#spring.profiles.active = prod
