--- src/main/resources/backend.properties
@@ -1,1 +1,0 @@
-backend.profile = dev
