--- src/test/java/com/expeditors/animal/AbstractAnimalTest.java
@@ -1,30 +1,0 @@
-package com.expeditors.animal;
-
-import org.junit.jupiter.api.Test;
-
-class AbstractAnimalTest {
-
-    @Test
-    void getAnimalName() {
-    }
-
-    @Test
-    void getAdoptedDate() {
-    }
-
-    @Test
-    void animalGreeting() {
-    }
-
-    @Test
-    void getOwner() {
-    }
-
-    @Test
-    void testToString() {
-    }
-
-    @Test
-    void setOwner() {
-    }
-}
